Moved test suites to centralized location and a few fixes
-Moved all test suites and snapshots from their distributed locations into the src/tests/ folder. CRA configures Jest to only run tests within the 'src/' dir. For tests needed to be run outside of this folder, a separate jest config file will be required.
-Implemented a RTL wrapper to disable MUI's dynamic JSS classes generation and applied it to all rendering tests.
-Changed all the snapshots according to the renders produced by the wrapper.
-Exported the default 'exchangeRates' object from helpers.ts for its usage on its tests.
-Added a test for WalletWidget.tsx to check that 'walletSelect' runs upon clicking on wallet connection button.
