+++ src/tests/CurrencyInfoCard.test.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
-import CurrencyInfoCard from '../CurrencyInfoCard';
+import CurrencyInfoCard from '../components-v2/common/CurrencyInfoCard';
 import '@testing-library/jest-dom';
 import BigNumber from 'bignumber.js';
-import { render } from '@testing-library/react';
+import { customRender } from './Utils';
 
 describe('CurrencyInfoCard', () => {
 	test('Renders correctly', () => {
 		const displayPrice = new BigNumber(40);
 		const currency = 'usd';
-		const { container } = render(
+		const { container } = customRender(
 			<CurrencyInfoCard title="Test Price" value={displayPrice} currency={currency} isUsd={true} />,
 		);
 		expect(container).toMatchSnapshot();
